14 June: json read/write

+++ src/main/java/com/plc/json/repository/JsonRepository.java
@@ -3,7 +3,7 @@ package com.plc.json.repository;
 import com.plc.json.model.Jsondata;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JsonRepository extends JpaRepository<Jsondata,Long> {
+public interface JsonRepository extends JpaRepository<Jsondata,Integer> {
 
 
 }
